check

+++ src/mobile/data/Abonent.kt
@@ -1,5 +1,5 @@
 package mobile.data
-data class Abonent(val fullName: String, val phoneNumber: String, val balance: Float, val company: String){
+data class Abonent(val fullName: String, val phoneNumber: String, val balance: Double, val company: String){
 //    var operator: String = ""
 //    fun operator(){
 //        if (phoneNumber.startsWith("93") || phoneNumber.startsWith("94")) operator = "USELL"
